solution/api/requirements.txt
pandas==2.2.3
duckdb==1.1.1
plotly==5.24.1
Flask==3.0.2
SQLAlchemy==2.0.28
psycopg2-binary==2.9.9
python-dotenv==1.0.1